refactor: Dealer와 Player가 abstract Role을 상속받도록 수정
