+++ src/main/java/blackjack/domain/gamer/role/Role.java
@@ -1,36 +1,25 @@
-package blackjack.domain.gamer;
+package blackjack.domain.gamer.role;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import blackjack.domain.Money;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardGroup;
-import blackjack.domain.result.Match;
 
-public class Player{
+public abstract class Role {
     private static final String NAME_LENGTH_ERROR_MESSAGE = "이름은 최소 1글자, 최대 6글자로 입력해야 합니다.";
     private static final String NAME_EMPTY_ERROR_MESSAGE = "빈 문자는 이름으로 입력할 수 없습니다.";
     private static final int MINIMUM_NAME_LENGTH = 1;
     private static final int MAXIMUM_NAME_LENGTH = 6;
 
-    private final CardGroup cardGroup = new CardGroup();
+    protected final CardGroup cardGroup = new CardGroup();
     private final String name;
-    private final Money betMoney;
 
-    public Player(String name) {
+    public Role(String name) {
         validateName(name);
         this.name = name;
-        this.betMoney = new Money(0);
     }
 
-    public static List<Player> of(List<String> playerNames) {
-        List<Player> players = new ArrayList<>();
-        for (String playerName : playerNames) {
-            players.add(new Player(playerName));
-        }
-        return players;
-    }
+    abstract public boolean isAddable();
+
+    abstract public void addCard(Card card);
 
     private void validateName(String name) {
         validateEmpty(name);
@@ -49,25 +38,6 @@ public class Player{
         }
     }
 
-    public int compareCardsSumTo(int sum) {
-        if (isBust()) {
-            return -1;
-        }
-
-        if (isBust(sum)) {
-            return 1;
-        }
-
-        return Integer.compare(getScore(), sum);
-    }
-
-    public void addCard(Card card) {
-        if (cardGroup.isBust()) {
-            return;
-        }
-        cardGroup.addCard(card);
-    }
-
     public void openAllCards() {
         cardGroup.open();
     }
@@ -78,10 +48,6 @@ public class Player{
 
     public boolean isBust(int sum) {
         return cardGroup.isBust(sum);
-    }
-
-    public boolean isAddable() {
-        return cardGroup.isAddable();
     }
 
     public int getScore() {
@@ -98,26 +64,5 @@ public class Player{
 
     public int getCardsSize() {
         return cardGroup.getSize();
-    }
-
-    public void addMoney(int amount) {
-        betMoney.add(amount);
-    }
-
-    public int getProfitMoney(Match match) {
-        return (int) (betMoney.getAmount() * profitRatio(match));
-    }
-
-    private double profitRatio(Match match) {
-        if (match.equals(Match.WIN) && cardGroup.isBlackJack()) {
-            return 1.5;
-        }
-        if (match.equals(Match.WIN)) {
-            return 1;
-        }
-        if (match.equals(Match.LOSE)) {
-            return -1;
-        }
-        return 0;
     }
 }
