+++ src/main/java/blackjack/dto/GamerCardsResultDto.java
@@ -5,7 +5,7 @@ import java.util.Collections;
 import java.util.List;
 
 import blackjack.domain.card.CardGroup;
-import blackjack.domain.gamer.Player;
+import blackjack.domain.gamer.role.Role;
 
 public class GamerCardsResultDto {
     private final GamerCardsDto gamerCardsDto;
@@ -19,10 +19,10 @@ public class GamerCardsResultDto {
     public static GamerCardsResultDto of(String name, CardGroup cards, int sum) {
         return new GamerCardsResultDto(GamerCardsDto.of(name, cards), sum);
     }
-    public static List<GamerCardsResultDto> of(List<Player> playersCards) {
+    public static List<GamerCardsResultDto> of(List<Role> roles) {
         List<GamerCardsResultDto> gamersCardsResultDto = new ArrayList<>();
-        for (Player player : playersCards) {
-            gamersCardsResultDto.add(of(player.getName(), player.getCardGroup(), player.getScore()));
+        for (Role role : roles) {
+            gamersCardsResultDto.add(of(role.getName(), role.getCardGroup(), role.getScore()));
         }
         return Collections.unmodifiableList(gamersCardsResultDto);
     }
