+++ src/test/java/blackjack/domain/BlackJackTest.java
@@ -12,9 +12,9 @@ import org.junit.jupiter.api.Test;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.property.CardNumber;
 import blackjack.domain.card.property.CardShape;
-import blackjack.domain.gamer.Dealer;
-import blackjack.domain.gamer.Player;
 import blackjack.domain.gamer.PlayerGroup;
+import blackjack.domain.gamer.role.Dealer;
+import blackjack.domain.gamer.role.Player;
 import blackjack.domain.result.DealerResult;
 import blackjack.domain.result.GameResult;
 import blackjack.domain.result.Match;
