+++ src/test/java/blackjack/domain/PlayerGroupTest.java
@@ -1,15 +1,17 @@
 package blackjack.domain;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.*;
 
-import blackjack.domain.gamer.Player;
-import blackjack.domain.gamer.PlayerGroup;
 import java.util.List;
 import java.util.stream.Stream;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import blackjack.domain.gamer.PlayerGroup;
+import blackjack.domain.gamer.role.Player;
 
 class PlayerGroupTest {
