+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import blackjack.domain.gamer.Player;
+import blackjack.domain.gamer.role.Player;
 
 class PlayerTest {
